refactor: set all store in one place
